added exception handling
added try statements to catch exceptions.
put zero padding code in zero_padder function.
included check for file to prevent directories from being renamed.
pass filetypes to file_counter to prevent two calls to list_of_filetypes_modifier()

+++ picture_namer.py
@@ -28,21 +28,29 @@ import argparse
 
 
 def main():
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        '-f', '--filename', help='Sort by filename', action='store_true')
-    parser.add_argument(
-        '-m', '--moddate', help='Sort by file modified date',
-        action='store_true')
-    parser.add_argument(
-        '-p', '--folderpath', type=str, help='Add a path to the folder')
-    args = parser.parse_args()
-    if args.folderpath:
-        path = args.folderpath
-    else:
-        path = input('Folder path: ')
-    file_namer(path, args.filename, args.moddate)
-    input('Press Enter to exit.')
+    try:
+        parser = argparse.ArgumentParser()
+        parser.add_argument(
+            '-f', '--filename', help='Sort by filename', action='store_true')
+        parser.add_argument(
+            '-m', '--moddate', help='Sort by file modified date',
+            action='store_true')
+        parser.add_argument(
+            '-p', '--folderpath', type=str, help='Add a path to the folder')
+        args = parser.parse_args()
+        if args.folderpath:
+            path = args.folderpath
+            if path.endswith('"'):  # remove " from end if present
+                path = path[:-1]
+        else:
+            path = input('Folder path to images: ')
+        file_namer(path, args.filename, args.moddate)
+    except Exception as e:
+        print(e)
+    try:
+        input('Press Enter to exit...')
+    except Exception as e:
+        print(e)
     return 0
 
 
@@ -70,7 +78,7 @@ def get_filetypes():
         else:
             print('Error: _list_of_filetypes.txt not found.')
             path = input('Input path with txt file containing filetypes: ')
-        if (os.path.isfile(path)):
+        if os.path.isfile(path):
             filetypes_file = open(path, 'r')
             list_of_filetypes = filetypes_file.readlines()
             filetypes_file.close()
@@ -90,10 +98,10 @@ def clean_list_of_filetypes(list_of_filetypes):
         filetype = filetype.replace(' ', '')
         comment_c = filetype.find('//')
         comment_py = filetype.find('#')
-        if (comment_c >= 0 or comment_py >= 0):
-            if (comment_c == 0 or comment_py == 0):
+        if comment_c >= 0 or comment_py >= 0:
+            if comment_c == 0 or comment_py == 0:
                 continue
-            elif (comment_c < comment_py or comment_py == (-1)):
+            elif comment_c < comment_py or comment_py == (-1):
                 clean_list_of_filetypes.append(filetype[:comment_c])
             else:
                 clean_list_of_filetypes.append(filetype[:comment_py])
@@ -102,13 +110,22 @@ def clean_list_of_filetypes(list_of_filetypes):
     return clean_list_of_filetypes
 
 
-def file_namer(path, argfilename, argmoddate):
-    count = file_counter()
-    lead_zeros = 5
+def zero_padder(lead_zeros, count):
     if len(str(count)) >= lead_zeros:
         lead_zeros = len(str(count)) + 2
+    return lead_zeros
+
+
+def file_namer(path, argfilename, argmoddate):
+    files_renamed = 0
     filetypes = list_of_filetypes_modifier()
-    os.chdir(path)
+    try:
+        os.chdir(path)
+    except Exception as e:
+        print('Error with inputted path.', e, 'occurred.')
+        return
+    count = file_counter(filetypes)
+    lead_zeros = zero_padder(5, count)  # Ensures leading zeros
     files = os.listdir('.')
     if argfilename is True:
         files = sorted(files)
@@ -117,28 +134,35 @@ def file_namer(path, argfilename, argmoddate):
     else:
         files = sorted(files, key=os.path.getmtime)
     for file in files:
-        if file.endswith(tuple(filetypes)):
-            rel_path = os.path.relpath('.', '..').replace(' ', '_')
-            if file.startswith(rel_path):
-                continue
-            else:
-                new_file = os.path.relpath('.', '..') + '_'
-                new_file = new_file.replace(' ', '_')
-                new_file = new_file + str(count).zfill(lead_zeros) + '_'
-                print(file + ' -> ' + new_file + file)
-                os.rename(file, new_file + file)
-                count = count + 1
-    print('Renamed', count, 'files.')
+        if os.path.isfile(file):
+            if file.endswith(tuple(filetypes)):
+                rel_path = os.path.relpath('.', '..').replace(' ', '_')
+                if file.startswith(rel_path):
+                    continue
+                else:
+                    new_file = os.path.relpath('.', '..') + '_'
+                    new_file = new_file.replace(' ', '_')
+                    new_file = new_file + str(count).zfill(lead_zeros) + '_'
+                    print(file + ' -> ' + new_file + file)
+                    os.rename(file, new_file + file)
+                    count = count + 1
+                    files_renamed = files_renamed + 1
+        else:
+            continue
+    print('Renamed', files_renamed, 'files.')
+    return
 
 
-def file_counter():  # returns the number of files in a directory
-    filetypes = list_of_filetypes_modifier()
+def file_counter(filetypes):  # returns the number of files in a directory
     count = 0
     for file in os.listdir('.'):
-        if file.endswith(tuple(filetypes)):
-            rel_path = os.path.relpath('.', '..').replace(' ', '_')
-            if file.startswith(rel_path):
-                count = count + 1
+        if os.path.isfile(file):
+            if file.endswith(tuple(filetypes)):
+                rel_path = os.path.relpath('.', '..').replace(' ', '_')
+                if file.startswith(rel_path):
+                    count = count + 1
+        else:
+            continue
     return count
